Placed each earth model example in seperate file (#30)
