+++ doc/api/index.rst
@@ -15,6 +15,15 @@ Data Fetching and Loading
 
     fetch_etopo1
     fetch_prem
+    fetch_ak135f
+    fetch_iasp91
+    fetch_mean
+    fetch_pema
+    fetch_pemc
+    fetch_pemo
+    fetch_mc35
+    fetch_stw105
+    fetch_tna_sna
     fetch_bedmap2
     fetch_seafloor_age
